+++ persistent/ChangeLog.md
@@ -5,6 +5,16 @@
 * [#1440](https://github.com/yesodweb/persistent/pull/1440)
     * Defined NFData PersistValue
 
+## 2.14.3.2
+
+* [#1446](https://github.com/yesodweb/persistent/pull/1446)
+    * Foreign key discovery was fixed for qualified names, `Key Model`, and
+      `Maybe` references.
+* [#1438](https://github.com/yesodweb/persistent/pull/1438)
+    * Clarify wording on the error message for null in unique constraint
+* [#1447](https://github.com/yesodweb/persistent/pull/1447)
+    * Fix `SafeToInsert` not being generated correctly for some `Id` columns
+
 ## 2.14.3.1
 
 * [#1428](https://github.com/yesodweb/persistent/pull/1428)